requirements.txt
supabase
pandas
google-generativeai
requests
python-dotenv
